added & updated more tests
